게시물 수정, 삭제 시 로그인 체크

+++ src/main/java/com/example/sbs/cuni/controller/ArticleController.java
@@ -3,6 +3,8 @@ package com.example.sbs.cuni.controller;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -39,11 +41,32 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("article/doDelete")
-	public String doDelete(Model model, int id) {
+	public String doDelete(Model model, int id, HttpSession session) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirct";
+		}
+		
+		Article article = articleService.getArticle(id);
+		
+		int boardId = article.getBoardId();
+		
+		Board board = articleService.getBoard(boardId);
+		
+		String boardCode = board.getCode();
+		
 		Map<String, Object> rs = articleService.deleteArticle(id);
 		
 		String msg = (String) rs.get("msg");
-		String redirectUrl = "/article/list";
+		String redirectUrl = "/article/list?boardCode=" + boardCode;
 		
 		model.addAttribute("alertMsg", msg);
 		model.addAttribute("locationReplace", redirectUrl);
@@ -52,7 +75,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("article/write")
-	public String showWrite(Model model, String boardCode) {
+	public String showWrite(Model model, String boardCode, HttpSession session) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirct";
+		}
+		
 		Board board = articleService.getBoard(boardCode);
 		
 		model.addAttribute("board", board);
@@ -61,7 +97,22 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("article/doWrite")
-	public String doWrite(Model model, @RequestParam Map<String, Object> param) {
+	public String doWrite(Model model, @RequestParam Map<String, Object> param, HttpSession session) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirct";
+		}
+		
+		param.put("memberId", loginedMemberId);
+		
 		Map<String, Object> rs = articleService.writeArticle(param);
 		
 		int boardId = Integer.parseInt((String) param.get("boardId"));
@@ -78,7 +129,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("article/modify")
-	public String showModify(Model model, int id) {
+	public String showModify(Model model, int id, HttpSession session) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirct";
+		}
+		
 		Article article = articleService.getArticle(id);
 		
 		model.addAttribute("article", article);
@@ -87,7 +151,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("article/doModify")
-	public String doModify(Model model, @RequestParam Map<String, Object> param) {
+	public String doModify(Model model, @RequestParam Map<String, Object> param, HttpSession session) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirct";
+		}
+		
 		Map<String, Object> rs = articleService.modifyArticle(param);
 		
 		int id = Integer.parseInt((String) param.get("id"));
